Implemented turn-based battle system
I implemented a battle based turn system. This is includes new tables to hold enemies and player stats. A new UI that shows enemy health, player's mana and experience gain from multiple fights. buttons are created to let the player choose to attack or cast a spell. spells cost mana to use and take multiple days to replenish. multiple images of enemies saved in the app. Once an enemy dies, a new enemy is replaced and the player gains experience points. The player can only attack after the user inputs their daily activity. a new button is added to the main menu to navigate to the battle screen.

in the battle screen, the player can navigate to the scenario screen to reread the scenario given for that day. New scenarios are implemented into the app.